Working MCTS with subtree replacement selection based on 1/sims

+++ src/AI/Genetic_Algorithm/Population_EvalFunc.java
@@ -229,7 +229,7 @@ public class Population_EvalFunc {
      * @return best performing AI
      */
     public AI getTopSpecimen() {
-        double fitness = Double.MIN_VALUE;
+        double fitness = Integer.MIN_VALUE;
         int indexTopSpecimen = -1;
         for(int i = 0; i < AIs.length; i++) {
             if(AIs[i].getFitness() > fitness) {
@@ -245,7 +245,7 @@ public class Population_EvalFunc {
      * @return worst performing AI
      */
     public AI getWorstSpecimen() {
-        double fitness = Double.MAX_VALUE;
+        double fitness = Integer.MAX_VALUE;
         int indexWorstSpecimen = -1;
         for(int i = 0; i < AIs.length; i++) {
             if(AIs[i].getFitness() < fitness) {
